requirements.txt
Django==1.11.6
python-decouple==3.1
dj-database-url==0.4.2
psycopg2==2.7.3.2
djangorestframework==3.8.2
markdown==2.6.11
django-filter==1.1.0